# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Application",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/computers": {
            "get": {
                "tags": [
                    "computers"
                ],
                "summary": "Get a list of computers",
                "description": "",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "computers"
                ],
                "summary": "Create new computers",
                "description": "",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name book",
                                        "type": "string"
                                    },
                                    "desc": {
                                        "description": "about book",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "about image",
                                        "type": "string"
                                    },
                                    "monofacture_id": {
                                        "description": "monofacture id",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Validate error"
                    },
                    "200": {
                        "description": "Success create element"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/computers/{id}": {
            "get": {
                "tags": [
                    "computers"
                ],
                "summary": "Get one computers with all atributes",
                "description": "",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Computer id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "computers"
                ],
                "summary": "Updates a computers in the store with form data",
                "description": "",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Computer that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name",
                                        "type": "string"
                                    },
                                    "desc": {
                                        "description": "desc",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "image",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Error in server"
                    },
                    "400": {
                        "description": "Validate error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}